--- ghost/api-version-compatibility-service/README.md
@@ -1,3 +1,0 @@
-# Api Version Compatibility Service
-
-Service to notify Ghost instance owners about API version compatibility issues
