--- ghost/api-version-compatibility-service/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/APIVersionCompatibilityService');
