+++ ghost/core/core/server/services/api-version-compatibility/templates/generic-mismatch.html
