+++ ghost/core/package.json
@@ -69,7 +69,6 @@
     "@tryghost/admin-api-schema": "4.5.5",
     "@tryghost/announcement-bar-settings": "0.0.0",
     "@tryghost/api-framework": "0.0.0",
-    "@tryghost/api-version-compatibility-service": "0.0.0",
     "@tryghost/bookshelf-plugins": "0.6.25",
     "@tryghost/color-utils": "0.2.2",
     "@tryghost/config-url-helpers": "1.0.12",
